refactor: Update import statements for mj-service to use mjServiceHelper
